requirements.txt
langgraph 
langchain 
python-dotenv 
groq 
langchain-groq 
openai 
langsmith 
langchain-openai 
langfuse 
openlit 
crewai 
crewai-tools 
sentence-transformers 
beautifulsoup4 
langchain-community 
faiss-cpu 
langchain-mcp-adapters 
fastapi 
nest-asyncio 
pyngrok 
uvicorn 
PyMuPDF 
python-docx